+++ test/qunit/test.js
@@ -1,17 +1,14 @@
 /*globals test, asyncTest, ConfigLoader, start, ok*/
+/* jshint devel:true */
 test("exported", function() {
 	ok(ConfigLoader, "Object Exported");
 });
-
 
 asyncTest("config loader", 7, function() {
 	var cl = new ConfigLoader({
 		uri: "mgid:cms:video:nickjr.com:119998",
 		configParams: {
 			someConfigParam: "param1"
-		},
-		mediaGenParams: {
-			someMediaGenParam: "param2"
 		}
 	});
 	cl.on(ConfigLoader.Events.READY, function(event) {
@@ -29,6 +26,49 @@ asyncTest("config loader", 7, function() {
 	cl.load();
 });
 
+asyncTest("media gen property", 7, function() {
+	var cl = new ConfigLoader({
+		uri: "mgid:uma:videolist:mtv.com:1712673",
+		mediaGenProperty: "brightcove_mediagenRootURL",
+		mediaGenURL: "data/mediaGen.json",
+		configParams: {
+			ref: "http://media.mtvnservices.com/player/api/xbox/MTV_App_XBoxone_v1",
+			someConfigParam: "param1"
+		}
+	});
+	cl.on(ConfigLoader.Events.READY, function(event) {
+		// equal totally causes the tests to hang :(
+		ok(event.type === ConfigLoader.Events.READY, "event type READY");
+		ok(event.target === cl, "event target match");
+		var config = event.data;
+		ok(config, "config exists");
+		ok(config.mediaGen, "config.mediaGen exists");
+		ok(config.mediaGen.vmap, "config.mediaGen.vmap exists");
+		ok(config.mediaGen.vmap.adBreaks, "config.mediaGen.vmap.adBreaks exists");
+		ok(config.mediaGen.vmap.trackers, "config.mediaGen.vmap.trackers exists");
+		console.log("test.js:49 config", config);
+		start();
+	});
+	cl.on(ConfigLoader.Events.ERROR, function(event) {
+		console.log("test.js:49 event", event.data);
+		ok(false, "error thrown");
+		start();
+	});
+	cl.load();
+});
+
+asyncTest("json parse error", 1, function() {
+	var cl = new ConfigLoader({
+		configURL: "http://media.mtvnservices-q.mtvi.com/pmt/e1/access/index.html?returntype=config&configtype=html&stage=d"
+	});
+	cl.on(ConfigLoader.Events.ERROR, function(event) {
+		console.log("test.js:49 event", event.data);
+		ok(true, "error thrown");
+		start();
+	});
+	cl.load();
+});
+
 asyncTest("test error", 2, function() {
 	var cl = new ConfigLoader({
 		configURL: "http://google.com",
@@ -36,11 +76,27 @@ asyncTest("test error", 2, function() {
 	});
 	cl.on(ConfigLoader.Events.ERROR, function(event) {
 		// equal totally causes the tests to hang :(
-		ok(event.type === ConfigLoader.Events.ERROR);
-		ok(event.target === cl, "event target match");
+		ok(event.type === ConfigLoader.Events.ERROR, "async test error, type");
+		ok(event.target === cl, "async test error, event target match");
 		start();
 	});
 	cl.load();
+});
+
+asyncTest("test mediaGen parser error", 2, function() {
+	var cl = new ConfigLoader({
+		uri: "mgid:cms:video:nickjr.com:119998"
+	});
+	cl.on(ConfigLoader.Events.ERROR, function(event) {
+		console.log("test mediaGen parser error:", event.data);
+		// equal totally causes the tests to hang :(
+		ok(event.type === ConfigLoader.Events.ERROR, "event type");
+		ok(event.target === cl, "event target match");
+		start();
+	});
+	setTimeout(function() {
+		cl.onMediaGenLoaded(undefined);
+	}, 500);
 });
 
 asyncTest("test abort", 1, function() {